updates to misc/update-pkg.sh

+++ misc/update-pkg.sh
@@ -14,11 +14,23 @@ die() {
 	exit 1
 }
 
+force=no
+if [ "x$1" = "x-f" ]; then
+	shift
+	force=yes
+fi
+
+keep=no
+if [ "x$1" = "x-k" ]; then
+	shift
+	keep=yes
+fi
+
 repo=$1
 shift
 
 case "$repo" in
-	core|extra|community) : ;;
+	core|extra|community|testing) : ;;
 	*) usage ;;
 esac
 
@@ -27,12 +39,15 @@ updatepkg() {
 	pkgfile=$1
 	pkgname=`echo "$pkgfile" | pcregrep -o '^.*?(?=-[0-9])'`
 
-	test -e "${target}/${repo}.db.tar.gz" || die "no ${repo}.db.tar.gz found in target directory"
+	if [ "x$force" = "xno" ]; then
+		test -e "${target}/${repo}.db.tar.gz" || die "no ${repo}.db.tar.gz found in target directory"
+	fi
 
 	test -e "$pkgfile" || die "$prog: $pkgfile: no such file"
 
 	case "$pkgfile" in
 		*-${CARCH}.pkg.tar.xz) : ;;
+		*-any.pkg.tar.xz) : ;;
 		*) die "invalid architecture" ;;
 	esac
 
@@ -53,7 +68,7 @@ updatepkg() {
 		esac
 	done
 	cd $olddir
-	rm -f "${pkgfile}" "${pkgfile}.sig"
+	[ "x$keep" = "xno" ] && rm -f "${pkgfile}" "${pkgfile}.sig"
 }
 
 for i in $@; do
